requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
brainflow>=5.10.0
sounddevice>=0.4.6
scipy>=1.10.0
python-dotenv>=1.0.0
